refactor: extract message publishing logic
